+++ server/src/main/scala/com/twitter/server/handler/LoggingHandler.scala
@@ -1,0 +1,8 @@
+package com.twitter.server.handler
+
+/** Marker trait for LoggingHandlers */
+trait LoggingHandler extends AdminHttpMuxHandler {
+
+  /** Implementation name */
+  def name: String
+}