+++ server/src/main/scala/com/twitter/server/handler/NoLoggingHandler.scala
@@ -9,10 +9,12 @@ import com.twitter.util.logging.Logger
 
 private[server] object NoLoggingHandler {
   val MissingLoggingImplMessageHeader: String =
-    "You have not configured a logging handler implementation for TwitterServer."
+    "You have not properly configured a logging handler implementation for TwitterServer."
   val MissingLoggingImplMessageBody: String =
-    "Please add a dependency on one of: twitter-server-logback-classic, " +
-      "twitter-server-slf4j-jdk14, or twitter-server-slf4j-log4j12."
+    "Please add a dependency on only one of: twitter-server-logback-classic, " +
+      "twitter-server-slf4j-jdk14, or twitter-server-slf4j-log4j12. Note: this page will appear " +
+      "when there is no dependency configured or when multiple conflicting dependencies have been " +
+      "detected. Please check the /admin/lint page for more information."
 }
 
 class NoLoggingHandler extends Service[Request, Response] {